Updated randomising weapons, removed floor_multiplier from item

+++ server/classes/armour.py
@@ -5,5 +5,5 @@ from pydantic import (
 from .item import Item
 
 class Armour(Item):
-    defence_ammount: int # How much defence this armour gives
-    max_health_increase: int # How much extra max health this armour gives
+    defence_ammount: float # How much defence this armour gives
+    max_health_increase: float # How much extra max health this armour gives
